Title: A/B 출력, Time: 8ms, Memory: 0MB, Status: Wrong Answer - Codetree

+++ 250903/A/B 출력/a-divide-b.cpp
@@ -6,7 +6,7 @@ int main() {
     // Please write your code here.
     int a, b;
     cin >> a >> b;
-    double c = a/b;
+    long double c = a/b;
     cout << fixed << setprecision(21);
     cout << c;
     return 0;
